Managed to reduce the code duplication by shifting the builder to densenet.py
